Split code of `restore.go` (#1384)
